+++ src/userinterface/view/PassengerLayout.fxml
@@ -16,7 +16,7 @@
 <?import javafx.scene.control.ChoiceBox?>
 <?import java.lang.String?>
 
-<BorderPane prefHeight="840.0" prefWidth="1024.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="userinterface.StaffController">
+<BorderPane prefHeight="840.0" prefWidth="1024.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="userinterface.ManagerController">
     <top>
         <SplitPane dividerPositions="0.5" prefHeight="40.0" prefWidth="200.0" BorderPane.alignment="CENTER">
             <items>
